FO-1765 Flyttet komponenter ut til separate filer
