+++ js/sider/InfoSideFO.js
@@ -5,15 +5,13 @@ import { bindActionCreators } from 'redux';
 import { getLedetekst } from 'digisyfo-npm';
 import Feilmelding from '../components/Feilmelding';
 import AppSpinner from '../components/AppSpinner';
-import ArbeidsrettetOppfolging from '../components/ArbeidsrettetOppfolging';
+import ArbeidsrettetOppfolging from '../components/arbeidsrettet-oppfolging/ArbeidsrettetOppfolging';
 import SideStrippet from './SideStrippet';
 import { hentOppfolging } from '../actions/brukerinfo_actions';
 
 class InfoSideFO extends Component {
     componentWillMount() {
-        if (!this.props.hentetOppfolging) {
-            this.props.actions.hentOppfolging();
-        }
+        this.props.actions.hentOppfolging();
     }
     render() {
         const {
@@ -45,7 +43,6 @@ InfoSideFO.propTypes = {
     hentingLedeteksterFeilet: PropTypes.bool,
     henterOppfolging: PropTypes.bool,
     hentingOppfolgingFeilet: PropTypes.bool,
-    hentetOppfolging: PropTypes.bool,
     underOppfolging: PropTypes.bool,
     actions: PropTypes.shape({
         hentOppfolging: PropTypes.func,
@@ -58,7 +55,6 @@ export function mapStateToProps(state) {
         hentingLedeteksterFeilet: state.ledetekster.hentingFeilet,
         henterOppfolging: state.brukerinfo.oppfolging.henter,
         hentingOppfolgingFeilet: state.brukerinfo.oppfolging.hentingFeilet,
-        hentetOppfolging: state.brukerinfo.oppfolging.hentet,
         underOppfolging: state.brukerinfo.oppfolging.data.underOppfolging,
     };
 }
